remove classes and models of outdated API version 1
