+++ lib/wanikani/models/study_material.rb
@@ -25,7 +25,7 @@ module Wanikani
       end
 
       def client
-        @client ||= ::Wanikani::ApiV2::Client.new(::Wanikani.config.to_hash)
+        @client ||= ::Wanikani::Client.new(::Wanikani.config.to_hash)
       end
     end
   end
